rename types in new reader
