+++ midix/src/parser/reader/mod.rs
@@ -4,10 +4,14 @@ mod error;
 pub use error::*;
 mod event;
 pub use event::*;
-mod header;
-pub use header::*;
-mod track;
-pub use track::*;
+mod header_chunk;
+pub use header_chunk::*;
+mod track_chunk;
+pub use track_chunk::*;
+mod track_event;
+pub use track_event::*;
+mod track_message;
+pub use track_message::*;
 
 use std::io::{BufRead, BufReader, Read};
